Add symbol table
The symbol table builds from a text description of the state of the
workflow. It includes types and restrictions on values. The syntax
is defined in an ANTLR grammar. The symbol table itself is built after
parsing the text desctription of the state with a listner using the
ANTLR runtime. Building a symbol table includes semantic checks:

  * No multiple definitions
  * If values are restricted for a variable, then the initial value
    must be included in the restricted set
  * Initial values for const types must be assignment compatable with
    the declared type
  * Initial values and restricted values for variables must be
    assignment compatable with the declared type

The commit includes the base machinary for type checking.