Reconfigure type checking and move old code
The project and pre-commit settings for `mypy` now use strict and ignore
files related to ANTLR. Specifically `mypy` for pre-commit ignores
anything from the old code and from ANTLR. Any missing type annotations
have been added to the new code to make `mypy` happy.

There is now an `__init__.py` to indicate the package. All the old code
is in the `original` folder. That should be deleted at some point.