Refactor state to use the builder and error field (#83)
Refactor to use a builder in conjunction with the listener to build an
instance of a `State` (formerly `SymbolTable`). The refactor touches a
whole bunch of files since the `State` is used in several tests and the
`cwp_ltl_visitor.py` file. That file will be refactored to use a public
interface in the next PR.

@tklaback, the visitor adds specific elements to the builder, and when
the listener finishes the traversal, it calls the
`state_builder.build()` to create an instance of the `State` object. All
of the error detection is in the `State` class and is triggered in the
build.

@jminyoon, the exception no longer carries the error; rather, I have a
`Maybe[Error]` attribute in the listener so that when the exception
fires, the error is taken from the listener rather than the exception.

Close #79
Fix #63